+++ instrumentation/opentelemetry-instrumentation-aiohttp-client/tests/test_aiohttp_client_integration.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# pylint: disable=too-many-lines
+
 import asyncio
 import contextlib
 import typing
@@ -88,7 +90,7 @@ class TestAioHttpIntegration(TestBase):
         super().setUp()
         _OpenTelemetrySemanticConventionStability._initialized = False
 
-    def assert_spans(self, spans, num_spans=1):
+    def _assert_spans(self, spans, num_spans=1):
         finished_spans = self.memory_exporter.get_finished_spans()
         self.assertEqual(num_spans, len(finished_spans))
         self.assertEqual(
@@ -102,6 +104,11 @@ class TestAioHttpIntegration(TestBase):
             ],
             spans,
         )
+
+    def _assert_metrics(self, num_metrics: int = 1):
+        metrics = self.get_sorted_metrics()
+        self.assertEqual(len(metrics), num_metrics)
+        return metrics
 
     @staticmethod
     def _http_request(
@@ -130,6 +137,7 @@ class TestAioHttpIntegration(TestBase):
         return run_with_test_server(client_request, url, handler)
 
     def test_status_codes(self):
+        index = 0
         for status_code, span_status in self._test_status_codes:
             with self.subTest(status_code=status_code):
                 path = "test-path?query=param#foobar"
@@ -145,10 +153,21 @@ class TestAioHttpIntegration(TestBase):
                     HTTP_STATUS_CODE: status_code,
                 }
                 spans = [("GET", (span_status, None), attributes)]
-                self.assert_spans(spans)
+                self._assert_spans(spans)
                 self.memory_exporter.clear()
+                metrics = self._assert_metrics(1)
+                duration_data_point = metrics[0].data.data_points[index]
+                self.assertEqual(
+                    dict(duration_data_point.attributes),
+                    {
+                        HTTP_STATUS_CODE: status_code,
+                        HTTP_METHOD: "GET",
+                    },
+                )
+                index += 1
 
     def test_status_codes_new_semconv(self):
+        index = 0
         for status_code, span_status in self._test_status_codes:
             with self.subTest(status_code=status_code):
                 path = "test-path?query=param#foobar"
@@ -168,10 +187,29 @@ class TestAioHttpIntegration(TestBase):
                 if status_code >= 400:
                     attributes[ERROR_TYPE] = str(status_code.value)
                 spans = [("GET", (span_status, None), attributes)]
-                self.assert_spans(spans)
+                self._assert_spans(spans)
                 self.memory_exporter.clear()
+                metrics = self._assert_metrics(1)
+                duration_data_point = metrics[0].data.data_points[index]
+                self.assertEqual(
+                    duration_data_point.attributes.get(
+                        HTTP_RESPONSE_STATUS_CODE
+                    ),
+                    status_code,
+                )
+                self.assertEqual(
+                    duration_data_point.attributes.get(HTTP_REQUEST_METHOD),
+                    "GET",
+                )
+                if status_code >= 400:
+                    self.assertEqual(
+                        duration_data_point.attributes.get(ERROR_TYPE),
+                        str(status_code.value),
+                    )
+                index += 1
 
     def test_status_codes_both_semconv(self):
+        index = 0
         for status_code, span_status in self._test_status_codes:
             with self.subTest(status_code=status_code):
                 path = "test-path?query=param#foobar"
@@ -195,8 +233,65 @@ class TestAioHttpIntegration(TestBase):
                     attributes[ERROR_TYPE] = str(status_code.value)
 
                 spans = [("GET", (span_status, None), attributes)]
-                self.assert_spans(spans, 1)
+                self._assert_spans(spans, 1)
                 self.memory_exporter.clear()
+                metrics = self._assert_metrics(2)
+                duration_data_point = metrics[0].data.data_points[index]
+                self.assertEqual(
+                    duration_data_point.attributes.get(HTTP_STATUS_CODE),
+                    status_code,
+                )
+                self.assertEqual(
+                    duration_data_point.attributes.get(HTTP_METHOD),
+                    "GET",
+                )
+                self.assertEqual(
+                    duration_data_point.attributes.get(ERROR_TYPE),
+                    None,
+                )
+                duration_data_point = metrics[1].data.data_points[index]
+                self.assertEqual(
+                    duration_data_point.attributes.get(
+                        HTTP_RESPONSE_STATUS_CODE
+                    ),
+                    status_code,
+                )
+                self.assertEqual(
+                    duration_data_point.attributes.get(HTTP_REQUEST_METHOD),
+                    "GET",
+                )
+                if status_code >= 400:
+                    self.assertEqual(
+                        duration_data_point.attributes.get(ERROR_TYPE),
+                        str(status_code.value),
+                    )
+                index += 1
+
+    def test_metrics(self):
+        with self.subTest(status_code=200):
+            # Run multiple requests to test metrics
+            number_of_requests = 3
+            for _ in range(number_of_requests):
+                self._http_request(
+                    trace_config=aiohttp_client.create_trace_config(),
+                    url="/test-path?query=param#foobar",
+                    status_code=200,
+                )
+
+            metrics = self._assert_metrics(1)
+            self.assertEqual(len(metrics[0].data.data_points), 1)
+            duration_data_point = metrics[0].data.data_points[0]
+            self.assertEqual(
+                dict(duration_data_point.attributes),
+                {
+                    HTTP_STATUS_CODE: 200,
+                    HTTP_METHOD: "GET",
+                },
+            )
+            self.assertEqual(duration_data_point.count, number_of_requests)
+            self.assertTrue(duration_data_point.min > 0)
+            self.assertTrue(duration_data_point.max > 0)
+            self.assertTrue(duration_data_point.sum > 0)
 
     def test_schema_url(self):
         with self.subTest(status_code=200):
@@ -319,7 +414,7 @@ class TestAioHttpIntegration(TestBase):
             status_code=HTTPStatus.OK,
         )
 
-        self.assert_spans(
+        self._assert_spans(
             [
                 (
                     "GET",
@@ -353,7 +448,7 @@ class TestAioHttpIntegration(TestBase):
                 with self.assertRaises(aiohttp.ClientConnectorError):
                     loop.run_until_complete(do_request(url))
 
-            self.assert_spans(
+            self._assert_spans(
                 [
                     (
                         "GET",
@@ -379,7 +474,7 @@ class TestAioHttpIntegration(TestBase):
         span = self.memory_exporter.get_finished_spans()[0]
         self.assertEqual(len(span.events), 1)
         self.assertEqual(span.events[0].name, "exception")
-        self.assert_spans(
+        self._assert_spans(
             [
                 (
                     "GET",
@@ -390,6 +485,14 @@ class TestAioHttpIntegration(TestBase):
                     },
                 )
             ]
+        )
+        metrics = self._assert_metrics(1)
+        duration_data_point = metrics[0].data.data_points[0]
+        self.assertEqual(
+            dict(duration_data_point.attributes),
+            {
+                HTTP_METHOD: "GET",
+            },
         )
 
     def test_basic_exception_new_semconv(self):
@@ -406,7 +509,7 @@ class TestAioHttpIntegration(TestBase):
         span = self.memory_exporter.get_finished_spans()[0]
         self.assertEqual(len(span.events), 1)
         self.assertEqual(span.events[0].name, "exception")
-        self.assert_spans(
+        self._assert_spans(
             [
                 (
                     "GET",
@@ -418,6 +521,15 @@ class TestAioHttpIntegration(TestBase):
                     },
                 )
             ]
+        )
+        metrics = self._assert_metrics(1)
+        duration_data_point = metrics[0].data.data_points[0]
+        self.assertEqual(
+            dict(duration_data_point.attributes),
+            {
+                HTTP_REQUEST_METHOD: "GET",
+                ERROR_TYPE: "ServerDisconnectedError",
+            },
         )
 
     def test_basic_exception_both_semconv(self):
@@ -434,7 +546,7 @@ class TestAioHttpIntegration(TestBase):
         span = self.memory_exporter.get_finished_spans()[0]
         self.assertEqual(len(span.events), 1)
         self.assertEqual(span.events[0].name, "exception")
-        self.assert_spans(
+        self._assert_spans(
             [
                 (
                     "GET",
@@ -448,6 +560,22 @@ class TestAioHttpIntegration(TestBase):
                     },
                 )
             ]
+        )
+        metrics = self._assert_metrics(2)
+        duration_data_point = metrics[0].data.data_points[0]
+        self.assertEqual(
+            dict(duration_data_point.attributes),
+            {
+                HTTP_METHOD: "GET",
+            },
+        )
+        duration_data_point = metrics[1].data.data_points[0]
+        self.assertEqual(
+            dict(duration_data_point.attributes),
+            {
+                HTTP_REQUEST_METHOD: "GET",
+                ERROR_TYPE: "ServerDisconnectedError",
+            },
         )
 
     def test_timeout(self):
@@ -463,7 +591,7 @@ class TestAioHttpIntegration(TestBase):
             timeout=aiohttp.ClientTimeout(sock_read=0.01),
         )
 
-        self.assert_spans(
+        self._assert_spans(
             [
                 (
                     "GET",
@@ -490,7 +618,7 @@ class TestAioHttpIntegration(TestBase):
             max_redirects=2,
         )
 
-        self.assert_spans(
+        self._assert_spans(
             [
                 (
                     "GET",
@@ -526,7 +654,7 @@ class TestAioHttpIntegration(TestBase):
                 loop = asyncio.get_event_loop()
                 loop.run_until_complete(do_request(url))
 
-        self.assert_spans(
+        self._assert_spans(
             [
                 (
                     "HTTP",
@@ -568,7 +696,7 @@ class TestAioHttpIntegration(TestBase):
                 loop = asyncio.get_event_loop()
                 loop.run_until_complete(do_request(url))
 
-        self.assert_spans(
+        self._assert_spans(
             [
                 (
                     "HTTP",
@@ -611,7 +739,7 @@ class TestAioHttpIntegration(TestBase):
                 loop = asyncio.get_event_loop()
                 loop.run_until_complete(do_request(url))
 
-        self.assert_spans(
+        self._assert_spans(
             [
                 (
                     "GET",
@@ -652,7 +780,7 @@ class TestAioHttpClientInstrumentor(TestBase):
 
         return default_request
 
-    def assert_spans(self, num_spans: int):
+    def _assert_spans(self, num_spans: int):
         finished_spans = self.memory_exporter.get_finished_spans()
         self.assertEqual(num_spans, len(finished_spans))
         if num_spans == 0:
@@ -661,11 +789,16 @@ class TestAioHttpClientInstrumentor(TestBase):
             return finished_spans[0]
         return finished_spans
 
+    def _assert_metrics(self, num_metrics: int = 1):
+        metrics = self.get_sorted_metrics()
+        self.assertEqual(len(metrics), num_metrics)
+        return metrics
+
     def test_instrument(self):
         host, port = run_with_test_server(
             self.get_default_request(), self.URL, self.default_handler
         )
-        span = self.assert_spans(1)
+        span = self._assert_spans(1)
         self.assertEqual("GET", span.name)
         self.assertEqual("GET", span.attributes[HTTP_METHOD])
         self.assertEqual(
@@ -673,6 +806,16 @@ class TestAioHttpClientInstrumentor(TestBase):
             span.attributes[HTTP_URL],
         )
         self.assertEqual(200, span.attributes[HTTP_STATUS_CODE])
+        metrics = self._assert_metrics(1)
+        duration_data_point = metrics[0].data.data_points[0]
+        self.assertEqual(duration_data_point.count, 1)
+        self.assertEqual(
+            dict(duration_data_point.attributes),
+            {
+                HTTP_STATUS_CODE: 200,
+                HTTP_METHOD: "GET",
+            },
+        )
 
     def test_instrument_new_semconv(self):
         AioHttpClientInstrumentor().uninstrument()
@@ -683,7 +826,7 @@ class TestAioHttpClientInstrumentor(TestBase):
             host, port = run_with_test_server(
                 self.get_default_request(), self.URL, self.default_handler
             )
-            span = self.assert_spans(1)
+            span = self._assert_spans(1)
             self.assertEqual("GET", span.name)
             self.assertEqual("GET", span.attributes[HTTP_REQUEST_METHOD])
             self.assertEqual(
@@ -691,6 +834,16 @@ class TestAioHttpClientInstrumentor(TestBase):
                 span.attributes[URL_FULL],
             )
             self.assertEqual(200, span.attributes[HTTP_RESPONSE_STATUS_CODE])
+            metrics = self._assert_metrics(1)
+            duration_data_point = metrics[0].data.data_points[0]
+            self.assertEqual(duration_data_point.count, 1)
+            self.assertEqual(
+                dict(duration_data_point.attributes),
+                {
+                    HTTP_RESPONSE_STATUS_CODE: 200,
+                    HTTP_REQUEST_METHOD: "GET",
+                },
+            )
 
     def test_instrument_both_semconv(self):
         AioHttpClientInstrumentor().uninstrument()
@@ -710,9 +863,28 @@ class TestAioHttpClientInstrumentor(TestBase):
                 HTTP_RESPONSE_STATUS_CODE: 200,
                 HTTP_STATUS_CODE: 200,
             }
-            span = self.assert_spans(1)
+            span = self._assert_spans(1)
             self.assertEqual("GET", span.name)
             self.assertEqual(span.attributes, attributes)
+            metrics = self._assert_metrics(2)
+            duration_data_point = metrics[0].data.data_points[0]
+            self.assertEqual(duration_data_point.count, 1)
+            self.assertEqual(
+                dict(duration_data_point.attributes),
+                {
+                    HTTP_STATUS_CODE: 200,
+                    HTTP_METHOD: "GET",
+                },
+            )
+            duration_data_point = metrics[1].data.data_points[0]
+            self.assertEqual(duration_data_point.count, 1)
+            self.assertEqual(
+                dict(duration_data_point.attributes),
+                {
+                    HTTP_RESPONSE_STATUS_CODE: 200,
+                    HTTP_REQUEST_METHOD: "GET",
+                },
+            )
 
     def test_instrument_with_custom_trace_config(self):
         trace_config = aiohttp.TraceConfig()
@@ -729,7 +901,7 @@ class TestAioHttpClientInstrumentor(TestBase):
                     await session.get(TestAioHttpClientInstrumentor.URL)
 
         run_with_test_server(make_request, self.URL, self.default_handler)
-        self.assert_spans(1)
+        self._assert_spans(1)
 
     def test_every_request_by_new_session_creates_one_span(self):
         async def make_request(server: aiohttp.test_utils.TestServer):
@@ -743,7 +915,7 @@ class TestAioHttpClientInstrumentor(TestBase):
                 run_with_test_server(
                     make_request, self.URL, self.default_handler
                 )
-                self.assert_spans(1)
+                self._assert_spans(1)
 
     def test_instrument_with_existing_trace_config(self):
         trace_config = aiohttp.TraceConfig()
@@ -760,7 +932,7 @@ class TestAioHttpClientInstrumentor(TestBase):
                     await session.get(TestAioHttpClientInstrumentor.URL)
 
         run_with_test_server(create_session, self.URL, self.default_handler)
-        self.assert_spans(1)
+        self._assert_spans(1)
 
     def test_no_op_tracer_provider(self):
         AioHttpClientInstrumentor().uninstrument()
@@ -780,13 +952,13 @@ class TestAioHttpClientInstrumentor(TestBase):
             self.get_default_request(), self.URL, self.default_handler
         )
 
-        self.assert_spans(0)
+        self._assert_spans(0)
 
         AioHttpClientInstrumentor().instrument()
         run_with_test_server(
             self.get_default_request(), self.URL, self.default_handler
         )
-        self.assert_spans(1)
+        self._assert_spans(1)
 
     def test_uninstrument_session(self):
         async def uninstrument_request(server: aiohttp.test_utils.TestServer):
@@ -798,19 +970,19 @@ class TestAioHttpClientInstrumentor(TestBase):
         run_with_test_server(
             uninstrument_request, self.URL, self.default_handler
         )
-        self.assert_spans(0)
+        self._assert_spans(0)
 
         run_with_test_server(
             self.get_default_request(), self.URL, self.default_handler
         )
-        self.assert_spans(1)
+        self._assert_spans(1)
 
     def test_suppress_instrumentation(self):
         with suppress_instrumentation():
             run_with_test_server(
                 self.get_default_request(), self.URL, self.default_handler
             )
-        self.assert_spans(0)
+        self._assert_spans(0)
 
     @staticmethod
     async def suppressed_request(server: aiohttp.test_utils.TestServer):
@@ -822,7 +994,7 @@ class TestAioHttpClientInstrumentor(TestBase):
         run_with_test_server(
             self.suppressed_request, self.URL, self.default_handler
         )
-        self.assert_spans(0)
+        self._assert_spans(0)
 
     def test_suppress_instrumentation_with_server_exception(self):
         # pylint:disable=unused-argument
@@ -832,7 +1004,7 @@ class TestAioHttpClientInstrumentor(TestBase):
         run_with_test_server(
             self.suppressed_request, self.URL, raising_handler
         )
-        self.assert_spans(0)
+        self._assert_spans(0)
 
     def test_url_filter(self):
         def strip_query_params(url: yarl.URL) -> str:
@@ -845,7 +1017,7 @@ class TestAioHttpClientInstrumentor(TestBase):
         host, port = run_with_test_server(
             self.get_default_request(url), url, self.default_handler
         )
-        span = self.assert_spans(1)
+        span = self._assert_spans(1)
         self.assertEqual(
             f"http://{host}:{port}/test-path",
             span.attributes[HTTP_URL],
@@ -873,7 +1045,7 @@ class TestAioHttpClientInstrumentor(TestBase):
         run_with_test_server(
             self.get_default_request(url), url, self.default_handler
         )
-        span = self.assert_spans(1)
+        span = self._assert_spans(1)
         self.assertEqual("GET - /test-path", span.name)
         self.assertIn("response_hook_attr", span.attributes)
         self.assertEqual(span.attributes["response_hook_attr"], "value")